refactor: factorize runner.Runner & Group in different files
